Disable non-ASCII paths for some more tests
